+++ Source/Runtime/Executor/Mono/MonoInputKeyCodeDownExecutor.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 
 namespace LazerLabs.Commands
 {
-    public abstract class MonoInputKeyCodeExecutor : DefaultMonoExecutor
+    public abstract class MonoInputKeyCodeDownExecutor : DefaultMonoExecutor
     {
         [SerializeField] private KeyCode m_inputKeyCode = default;
 
         protected virtual void Update()
         {
-            if (Input.GetKeyUp(m_inputKeyCode))
+            if (Input.GetKeyDown(m_inputKeyCode))
             {
                 Execute();
             }